Trying with bigdecimal

+++ src/main/java/com/futumap/webapi/dao/entity/JobEntity.java
@@ -1,6 +1,8 @@
 package com.futumap.webapi.dao.entity;
 
 import org.joda.time.DateTime;
+
+import java.math.BigDecimal;
 import java.util.Date;
 import javax.persistence.*;
 
@@ -30,10 +32,10 @@ public class JobEntity {
     private Double salary;
 
     @Column(name = "longitude")
-    private Double longitude;
+    private BigDecimal longitude;
 
     @Column(name = "latitude")
-    private Double latitude;
+    private BigDecimal latitude;
 
     @Column(name = "created_at")
     private Date createdAt;
@@ -45,7 +47,7 @@ public class JobEntity {
 
     }
 
-    public JobEntity(String title, String description, Double salary, Double longitude, Double latitude) {
+    public JobEntity(String title, String description, Double salary, BigDecimal longitude, BigDecimal latitude) {
         this.setTitle(title);
         this.setDescription(description);
         this.setSalary(salary);
@@ -94,19 +96,19 @@ public class JobEntity {
         this.updatedAt = new Date();
     }
 
-    public Double getLongitude() {
+    public BigDecimal getLongitude() {
         return longitude;
     }
 
-    private void setLongitude(Double longitude) {
+    private void setLongitude(BigDecimal longitude) {
         this.longitude = longitude;
     }
 
-    public Double getLatitude() {
+    public BigDecimal getLatitude() {
         return latitude;
     }
 
-    private void setLatitude(Double latitude) {
+    private void setLatitude(BigDecimal latitude) {
         this.latitude = latitude;
     }
