+++ src/main/java/com/futumap/webapi/respository/JobRepository.java
@@ -7,11 +7,12 @@ import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 import org.springframework.web.bind.annotation.RequestParam;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Repository
 public interface JobRepository extends JpaRepository<JobEntity, Integer> {
 
     @Query(value="SELECT * from jobs j WHERE earth_box(ll_to_earth(?1,?2),?3) @> ll_to_earth(j.longitude,j.latitude)", nativeQuery = true)
-    List<JobEntity> findNearestJobs(@Param("longitude") Double longitude, @Param("latitude") Double latitude, @Param("distance") Double distance);
+    List<JobEntity> findNearestJobs(@Param("longitude") BigDecimal longitude, @Param("latitude") BigDecimal latitude, @Param("distance") Double distance);
 }
