+++ src/main/java/com/futumap/webapi/service/IJobService.java
@@ -2,13 +2,14 @@ package com.futumap.webapi.service;
 
 import com.futumap.webapi.dao.entity.JobEntity;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
 public interface IJobService {
 
     List<JobEntity> findAll();
-    List<JobEntity> findNearestJobs(Double longitude, Double latitude, Double distance);
+    List<JobEntity> findNearestJobs(BigDecimal longitude, BigDecimal latitude, Double distance);
     JobEntity save (JobEntity jobEntity);
     JobEntity update(JobEntity jobEntity);
     void delete(Integer id);
