+++ src/main/java/com/futumap/webapi/service/JobService.java
@@ -5,6 +5,7 @@ import com.futumap.webapi.respository.JobRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
@@ -21,7 +22,7 @@ public class JobService implements IJobService {
 
 
     @Override
-    public List<JobEntity> findNearestJobs(Double longitude, Double latitude, Double distance) {
+    public List<JobEntity> findNearestJobs(BigDecimal longitude, BigDecimal latitude, Double distance) {
         return repository.findNearestJobs(longitude, latitude, distance);
     }
